changed dio_switch from uintptr_t to unsigned int

+++ include/PCI_init.h
@@ -71,7 +71,7 @@ struct pci_dev_info info;
 void *hdl;
 uintptr_t iobase[6];
 
-uintptr_t dio_switch;
+unsigned int dio_switch;
 int switch0;
 int switch0_prev;
 unsigned int i;
